refactor(blockifier): move TransactionResources to a separate file
