+++ crates/blockifier/src/bouncer_test.rs
@@ -184,7 +184,7 @@ fn test_bouncer_try_update(
 ) {
     use cairo_vm::vm::runners::cairo_runner::ExecutionResources;
 
-    use crate::transaction::objects::TransactionResources;
+    use crate::fee::resources::TransactionResources;
 
     let state = &mut test_state(&BlockContext::create_for_account_testing().chain_info, 0, &[]);
     let mut transactional_state = TransactionalState::create_transactional(state);
