+++ crates/blockifier/src/fee.rs
@@ -3,3 +3,4 @@ pub mod fee_checks;
 pub mod fee_utils;
 pub mod gas_usage;
 pub mod receipt;
+pub mod resources;
